refactor: replace AzureB2CProvider subclass with AzureProvider.from_b2c() factory
